+++ src/Completion/CompletionConditionInstanceRepository.php
@@ -1,10 +1,9 @@
 <?php
 
-
 namespace BristolSU\Support\Completion;
 
-
 use BristolSU\Support\Completion\Contracts\CompletionConditionInstanceRepository as CompletionConditionInstanceRepositoryContract;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 /**
  * Class to retrieve and change completion condition instances
@@ -38,5 +37,37 @@ class CompletionConditionInstanceRepository implements CompletionConditionInstan
     public function all()
     {
         return CompletionConditionInstance::all();
+    }
+
+    /**
+     * Get a completion condition instance by ID
+     * 
+     * @param int $id
+     * @return Contracts\CompletionConditionInstance
+     * 
+     * @throws ModelNotFoundException If the completion condition instance was not found
+     */
+    public function getById(int $id): \BristolSU\Support\Completion\Contracts\CompletionConditionInstance
+    {
+        return CompletionConditionInstance::findOrFail($id);
+    }
+
+    /**
+     * Update a completion condition instance
+     *
+     * Any number of parameters may be passed to be updated. The possible attributes are alias, name, description and settings
+     *
+     * @param int $id
+     * @param array $attributes
+     * @return Contracts\CompletionConditionInstance
+     * 
+     * @throws ModelNotFoundException If the completion condition instance was not found
+     */
+    public function update(int $id, array $attributes): \BristolSU\Support\Completion\Contracts\CompletionConditionInstance
+    {
+        $completionCondition = $this->getById($id);
+        $completionCondition->fill($attributes);
+        $completionCondition->save();
+        return $completionCondition;
     }
 }