+++ src/Filters/FilterServiceProvider.php
@@ -11,8 +11,10 @@ use BristolSU\Support\Filters\Contracts\FilterTester as FilterTesterContract;
 use BristolSU\Support\Filters\Filters\Group\GroupNameIs;
 use BristolSU\Support\Filters\Filters\Group\GroupTagged;
 use BristolSU\Support\Filters\Filters\Role\RoleHasPosition;
+use BristolSU\Support\Filters\Filters\Role\RoleTagged;
 use BristolSU\Support\Filters\Filters\User\UserEmailIs;
 use BristolSU\Support\Filters\Commands\CacheFilters;
+use BristolSU\Support\Filters\Filters\User\UserTagged;
 use Illuminate\Contracts\Cache\Repository as Cache;
 use Illuminate\Contracts\Container\BindingResolutionException;
 use Illuminate\Support\ServiceProvider;
@@ -52,11 +54,29 @@ class FilterServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        $this->app->make(FilterManagerContract::class)->register('group_tagged', GroupTagged::class);
-        $this->app->make(FilterManagerContract::class)->register('user_email_is', UserEmailIs::class);
-        $this->app->make(FilterManagerContract::class)->register('role_has_position', RoleHasPosition::class);
-        $this->app->make(FilterManagerContract::class)->register('group_name_is', GroupNameIs::class);
+        $this->app->call([$this, 'registerFilters']);
         $this->commands([CacheFilters::class]);
+    }
 
+    /**
+     * Register all filters the SDK provides by default.
+     *
+     * This method may be overridden to stop any default filters being loaded
+     *
+     * @param FilterManagerContract $filterManager
+     */
+    public function registerFilters(FilterManagerContract $filterManager)
+    {
+        // User Filters
+        $filterManager->register('user_email_is', UserEmailIs::class);
+        $filterManager->register('user_tagged', UserTagged::class);
+            
+        // Group Filters
+        $filterManager->register('group_name_is', GroupNameIs::class);
+        $filterManager->register('group_tagged', GroupTagged::class);
+        
+        // Role Filters
+        $filterManager->register('role_has_position', RoleHasPosition::class);
+        $filterManager->register('role_tagged', RoleTagged::class);
     }
 }