+++ tests/Completion/CompletionConditionInstanceRepositoryTest.php
@@ -72,4 +72,43 @@ class CompletionConditionInstanceRepositoryTest extends TestCase
         }
     }
     
+    /** @test */
+    public function getById_returns_a_completion_condition_instance_with_the_given_id(){
+        $ccI = factory(CompletionConditionInstance::class)->create();
+        factory(CompletionConditionInstance::class, 5)->create();
+        
+        $repository = new CompletionConditionInstanceRepository();
+        $resolvedCci = $repository->getById($ccI->id);
+        
+        $this->assertInstanceOf(CompletionConditionInstance::class, $resolvedCci);
+        $this->assertModelEquals($ccI, $resolvedCci);
+    }
+    
+    /** @test */
+    public function update_updates_the_completion_condition_instance(){
+        $ccI = factory(CompletionConditionInstance::class)->create([
+            'alias'=> 'OldAlias', 'name' => 'OldName', 'description' => 'OldDescription', 'settings' => ['val' => 'OldVal']
+        ]);
+        factory(CompletionConditionInstance::class, 5)->create();
+
+        $this->assertDatabaseHas('completion_condition_instances', [
+            'id' => $ccI->id, 'alias'=> 'OldAlias', 'name' => 'OldName', 'description' => 'OldDescription', 'settings' => json_encode(['val' => 'OldVal'])
+        ]);
+        
+        $repository = new CompletionConditionInstanceRepository();
+        $resolvedCci = $repository->update($ccI->id, [
+            'alias'=> 'NewAlias', 'name' => 'NewName', 'description' => 'NewDescription', 'settings' => ['val' => 'NewVal']
+        ]);
+
+        $this->assertDatabaseHas('completion_condition_instances', [
+            'id' => $ccI->id, 'alias'=> 'NewAlias', 'name' => 'NewName', 'description' => 'NewDescription', 'settings' => json_encode(['val' => 'NewVal'])
+        ]);
+        
+        $this->assertInstanceOf(CompletionConditionInstance::class, $resolvedCci);
+        $this->assertEquals($ccI->id, $resolvedCci->id);
+        $this->assertEquals('NewAlias', $resolvedCci->alias);
+        $this->assertEquals('NewName', $resolvedCci->name);
+        $this->assertEquals('NewDescription', $resolvedCci->description);
+        $this->assertEquals(['val' => 'NewVal'], $resolvedCci->settings);
+    }
 }